Refactor EF migration logic and minor code cleanups
- Moved EF Core migration logic to new DbContextExtensions.ApplyMigrations<TContext> with optional logging; HostExtensions now delegates to it.
- Made DeletedAtUtc property in IDeletableEntity interface-level only.
- Removed unnecessary casts in AesGcmStreamCipher and AesGcmStreamFormat.
- Fixed extern/static order in ShellLink ExtractIconEx declaration.
- Made fields private and explicit in StopwatchDebuggerTests.

+++ Sources/EasyExtensions.Crypto/AesGcmStreamCipher.cs
@@ -107,7 +107,7 @@ namespace EasyExtensions.Crypto
             {
                 throw new ArgumentOutOfRangeException(nameof(windowCap), "Window cap must be >= 4.");
             }
-            if (memoryLimitBytes.HasValue && memoryLimitBytes.Value < (long)(MinChunkSize * 4))
+            if (memoryLimitBytes.HasValue && memoryLimitBytes.Value < MinChunkSize * 4)
             {
                 throw new ArgumentOutOfRangeException(nameof(memoryLimitBytes), "Memory limit too small for safe operation.");
             }
